requirements.txt
streamlit
pandas
numpy
scipy
plotly
python-binance
